Refactor for second cop rule to be added.
