+++ sqlcop/checks/cross_join.py
